requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
openpyxl
google-genai==0.3.0
python-multipart